fix(start-plugin-core): address review on server-fn discovery crawl
- Enforce a path boundary in the crawl's root check so a sibling directory
  sharing a name prefix (e.g. `<root>-other`) is not treated as under the root.
  This also makes the separate-package e2e fixture exercise the provider-module
  phase as intended rather than matching via the loose prefix.
- Guard virtual modules on the raw id (`id.startsWith('\0')`) before `cleanId`
  strips the NUL prefix, matching the convention used elsewhere; the previous
  post-`cleanId` `includes('\0')` check never matched.
- Use curly braces on the two one-line `if` bodies, per AGENTS.md.
- Extract the crawl (`shouldWalkForServerFns` / `ensureServerModuleGraphTransformed`)
  into `server-fn-graph-crawl.ts` so the pure boundary logic can be unit-tested
  without importing the full plugin (which pulls in server-only entry imports).
- Add unit tests for `shouldWalkForServerFns` covering the sibling-prefix
  boundary, NUL-prefixed virtual ids, node_modules / non-transformable
  exclusions, and query-string handling.
